requirements.txt
word2number~=1.1
pyspellchecker~=0.6.2
tree_sitter~=0.19.0
GitPython~=3.1.24